repo: consolidate paper, reorganize packages, and remove artifacts
Move canonical LaTeX manuscript into analysis/paper and simplify its \graphicspath to prefer local paper figures.
Reorganize repository layout: move core models into src/lbo, create optimization/ and workflows/ subpackages, and relocate analysis entry scripts into analysis/scripts.
Remove noise and duplicated artifacts (*_fixed.py variants, extra requirements files, Makefile variants, venv/, __pycache__/, *.egg-info) and keep only output/manifest.json.
Update .gitignore, STRUCTURE.md and regenerate folder_tree.txt to reflect the new, reviewer-friendly structure.

This makes the repo cleaner, reproducible, and easier to install/test.

+++ analysis/scripts/evaluation_protocol.py
@@ -140,10 +140,10 @@ def calculate_bootstrap_metrics(data: pd.DataFrame) -> Dict[str, float]:
     coverage = np.mean(np.abs(prediction_errors) <= ci_width)
     
     return {
-        'accuracy': accuracy,
-        'coverage': coverage,
-        'mean_error': np.mean(np.abs(prediction_errors)),
-        'n_observations': n
+        'accuracy': float(accuracy),
+        'coverage': float(coverage),
+        'mean_error': float(np.mean(np.abs(prediction_errors))),
+        'n_observations': float(n)
     }
 
 def ablation_study(
